move tests into new dir
