+++ tox.ini
@@ -19,6 +19,7 @@ passenv = http_proxy https_proxy no_proxy SSL_CERT_FILE PYTEST_*
 deps =
   pip >= 19.1
 extras = testing
+changedir = {toxinidir}/tests
 commands = pytest \
            --cov "{envsitepackagesdir}/speet" \
            --cov-config "{toxinidir}/tox.ini" \
